--- proc-macros/.gitignore
@@ -1,3 +1,0 @@
-/target
-**/*.rs.bk
-Cargo.lock
