--- proc-macros/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "proc_macros"
-version = "0.0.0"
-authors = ["Zachary Edgell <zacharyedgell080005664@gmail.com>"]
-description = "A `cargo generate` template for quick-starting a procedural macro crate"
-keywords = ["template", "proc_macro", "procmacro"]
-edition = "2018"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-quote = "1"
-proc-macro2 = "1.0"
-syn = "1.0"
-
