+++ src/appcallback.rs
@@ -1,6 +1,9 @@
 use std::collections::HashMap;
+use tonic::{Code, Request, Response, Status};
 
 use crate::dapr::*;
+use crate::dapr::dapr::proto::{common, runtime};
+use crate::dapr::dapr::proto::runtime::v1::app_callback_server::AppCallback;
 
 /// InvokeRequest is the message to invoke a method with the data.
 pub type InvokeRequest = dapr::proto::common::v1::InvokeRequest;
@@ -61,4 +64,94 @@ impl TopicSubscription {
 
         topic_subscription
     }
+}
+
+pub struct AppCallbackService {
+    handlers: Vec<Handler>,
+}
+
+pub struct Handler {
+    pub pub_sub_name: String,
+    pub topic: String,
+    pub handler: Box<dyn HandlerMethod>,
+}
+
+#[tonic::async_trait]
+impl AppCallback for AppCallbackService {
+    async fn on_invoke(
+        &self,
+        _request: Request<common::v1::InvokeRequest>,
+    ) -> Result<Response<common::v1::InvokeResponse>, Status> {
+        Ok(Response::new(InvokeResponse::default()))
+    }
+
+    async fn list_topic_subscriptions(
+        &self,
+        _request: Request<()>,
+    ) -> Result<Response<runtime::v1::ListTopicSubscriptionsResponse>, Status> {
+        let topics = self
+            .handlers
+            .iter()
+            .fold(Vec::new(), |mut topics, handler| {
+                topics.push(TopicSubscription::new(
+                    handler.pub_sub_name.clone(),
+                    handler.topic.clone(),
+                    None,
+                ));
+                topics
+            });
+        Ok(Response::new(ListTopicSubscriptionsResponse {
+            subscriptions: topics,
+        }))
+    }
+
+    async fn on_topic_event(
+        &self,
+        request: Request<runtime::v1::TopicEventRequest>,
+    ) -> Result<Response<runtime::v1::TopicEventResponse>, Status> {
+        let request_inner = request.into_inner();
+        let pub_sub_name = request_inner.pubsub_name.clone();
+        let topic_name = request_inner.topic.clone();
+        let handler = self
+            .handlers
+            .iter()
+            .filter(|x| x.pub_sub_name == pub_sub_name && x.topic == topic_name)
+            .next();
+        if let Some(handler) = handler {
+            return handler.handler.handler(request_inner).await;
+        }
+        Err(Status::new(Code::Internal, "Handler Not Found"))
+    }
+
+    async fn list_input_bindings(
+        &self,
+        _request: Request<()>,
+    ) -> Result<Response<runtime::v1::ListInputBindingsResponse>, Status> {
+        Ok(Response::new(ListInputBindingsResponse::default()))
+    }
+
+    async fn on_binding_event(
+        &self,
+        _request: Request<runtime::v1::BindingEventRequest>,
+    ) -> Result<Response<runtime::v1::BindingEventResponse>, Status> {
+        Ok(Response::new(BindingEventResponse::default()))
+    }
+}
+
+impl AppCallbackService {
+    pub fn new() -> AppCallbackService {
+        AppCallbackService { handlers: vec![] }
+    }
+
+    pub fn add_handler(&mut self, handler: Handler) {
+        self.handlers.push(handler)
+    }
+}
+
+#[tonic::async_trait]
+pub trait HandlerMethod: Send + Sync + 'static {
+    async fn handler(
+        &self,
+        request: runtime::v1::TopicEventRequest,
+    ) -> Result<Response<runtime::v1::TopicEventResponse>, Status>;
 }
