+++ src/lib.rs
@@ -4,6 +4,10 @@ pub mod dapr;
 pub mod error;
 pub mod server;
 
+pub use serde;
+
+pub use serde_json;
+
 pub use client::Client;
 
 extern crate dapr_macros;
